Remove iroha_macro_traits crate
