+++ macro/Cargo.toml
@@ -19,8 +19,10 @@ maintenance = { status = "actively-developed" }
 [features]
 default = ["std"]
 std = []
-dump_decoded = ["iroha_derive/dump_decoded", "iroha_macro_traits/dump_decoded"]
+dump_decoded = ["iroha_derive/dump_decoded", "parity-scale-codec", "eyre", "once_cell"]
 
 [dependencies]
 iroha_derive = { version = "=2.0.0-pre-rc.3", path = "derive" }
-iroha_macro_traits = { version = "=2.0.0-pre-rc.3", path = "traits" }
+parity-scale-codec = { version = "2.3.1", optional = true }
+eyre = { verstion = "0.6.8", optional = true }
+once_cell = { version = "1.8.0", optional = true }
