--- macro/traits/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "iroha_macro_traits"
-version = "2.0.0-pre-rc.3"
-authors = ["Iroha 2 team <https://github.com/orgs/soramitsu/teams/iroha2>"]
-edition = "2021"
-
-[features]
-dump_decoded = ["parity-scale-codec", "eyre", "once_cell"]
-
-[dependencies]
-parity-scale-codec = { version = "2.3.1", optional = true }
-eyre = { verstion = "0.6.8", optional = true }
-once_cell = { version = "1.8.0", optional = true }
